Extract declaration parsing to own module
