+++ src/ast/expr.rs
@@ -144,12 +144,6 @@ impl From<String> for Name {
     }
 }
 
-impl Name {
-    pub fn new(s: String) -> Self {
-        Name(s)
-    }
-}
-
 impl Display for Name {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(f, "{}", self.0)
