+++ src/parser/statement.rs
@@ -1,144 +1,17 @@
-use std::sync::Arc;
 use std::vec;
-
-use miette::{NamedSource, SourceSpan};
 
 use crate::ast::expr::{Expr, ExprType, Literal};
 use crate::ast::stmt::{Stmt, StmtType};
 use crate::ast::token::{Token, TokenType};
 use crate::source_span_extensions::SourceSpanExtensions;
 
-use super::parser_error::ParserError::{self, *};
+use super::parser_error::ParserError::*;
 
 use super::macros::{check, consume, match_token};
 use super::{Parser, Result};
 
-struct InternalBlock {
-    stmts: Vec<Stmt>,
-    location: SourceSpan,
-    src: Arc<NamedSource<String>>,
-}
-
 impl Parser {
-    pub(super) fn stmt(&mut self) -> Result<Stmt> {
-        self.declaration()
-    }
-
-    fn declaration(&mut self) -> Result<Stmt> {
-        use TokenType::*;
-        match self.peek().token_type {
-            Var => self.var_declaration(),
-            Fun => self.fun_declaration(),
-            _ => self.statement(),
-        }
-    }
-
-    fn var_declaration(&mut self) -> Result<Stmt> {
-        use TokenType::*;
-        let var_location = self.advance().location;
-        let peek = self.peek();
-        if let Identifier(name) = &peek.token_type {
-            let name = name.clone();
-            self.advance();
-            let mut expr = None;
-            if match_token!(self, TokenType::Equal).is_some() {
-                expr = Some(self.expression()?)
-            }
-            let semicolon = consume!(self, Semicolon, |t| self.expected_semicolon(t));
-            Ok(Stmt::var(
-                name,
-                expr,
-                var_location.until(semicolon.location),
-                self.src.clone(),
-            ))
-        } else {
-            Err(ExpectedIdentifier {
-                src: self.src.clone(),
-                location: peek.location,
-            })
-        }
-    }
-
-    fn fun_declaration(&mut self) -> Result<Stmt> {
-        use TokenType::*;
-        let fun_location = self.advance().location;
-        let identifier = self.peek();
-        if let Identifier(name) = &identifier.token_type {
-            let name = name.clone();
-            self.advance();
-
-            let arguments = self.argument_list()?;
-
-            let left_brace = self.peek();
-            if !matches!(left_brace.token_type, LeftBrace) {
-                return Err(ExpectedLeftBrace {
-                    src: self.src.clone(),
-                    location: left_brace.location,
-                });
-            }
-
-            let body = self.block()?;
-
-            Ok(Stmt::function(
-                name,
-                arguments,
-                body.stmts,
-                fun_location.until(body.location),
-                self.src.clone(),
-            ))
-        } else {
-            Err(ExpectedIdentifier {
-                src: self.src.clone(),
-                location: identifier.location,
-            })
-        }
-    }
-
-    fn argument_list(&mut self) -> Result<Vec<String>> {
-        use TokenType::*;
-        consume!(self, LeftParen, |t: &Token| {
-            ExpectedLeftParen {
-                src: self.src.clone(),
-                location: self.previous_if_eof(t.location),
-            }
-        });
-        let mut arguments = vec![];
-        if !check!(self, RightParen) {
-            loop {
-                if arguments.len() >= 255 {
-                    self.errors.push(ParserError::TooManyArguments {
-                        src: self.peek().src.clone(),
-                        location: self.peek().location,
-                    })
-                }
-
-                let identifier = self.peek();
-                let identifier_location = identifier.location;
-                if let Identifier(arg_name) = &identifier.token_type {
-                    arguments.push(arg_name.clone());
-                    self.advance();
-                } else {
-                    return Err(ParserError::ExpectedIdentifier {
-                        src: identifier.src.clone(),
-                        location: identifier_location,
-                    });
-                }
-
-                if match_token!(self, Comma).is_none() {
-                    break;
-                }
-            }
-        }
-        consume!(self, RightParen, |t: &Token| {
-            ExpectedRightParen {
-                src: t.src.clone(),
-                location: self.previous_if_eof(t.location),
-            }
-        });
-        Ok(arguments)
-    }
-
-    fn statement(&mut self) -> Result<Stmt> {
+    pub(super) fn statement(&mut self) -> Result<Stmt> {
         use TokenType::*;
         match self.peek().token_type {
             Print => self.print_statement(),
@@ -299,27 +172,6 @@ impl Parser {
         Ok(while_statement)
     }
 
-    fn block(&mut self) -> Result<InternalBlock> {
-        let left_brace_location = self.advance().location;
-        let mut stmts = vec![];
-        while !check!(self, TokenType::RightBrace) && !self.is_at_end() {
-            stmts.push(self.declaration()?);
-        }
-
-        let right_brace = consume!(self, TokenType::RightBrace, |t: &Token| {
-            ExpectedRightBrace {
-                src: t.src.clone(),
-                location: self.previous_if_eof(t.location),
-            }
-        });
-
-        Ok(InternalBlock {
-            stmts,
-            location: left_brace_location.until(right_brace.location),
-            src: right_brace.src.clone(),
-        })
-    }
-
     fn expression_statement(&mut self) -> Result<Stmt> {
         let expr = self.expression()?;
         let semicolon = consume!(self, TokenType::Semicolon, |t: &Token| ExpectedSemicolon {
@@ -338,18 +190,10 @@ impl Parser {
         let location = print_token_location.until(semicolon.location);
         Ok(Stmt::print(expr, location))
     }
-
-    fn expected_semicolon(&self, t: &Token) -> ParserError {
-        ExpectedSemicolon {
-            expr: None,
-            src: t.src.clone(),
-            location: self.previous_if_eof(t.location),
-        }
-    }
 }
 
 #[cfg(test)]
-mod tests {
+mod test {
 
     use crate::{
         ast::{
@@ -363,7 +207,7 @@ mod tests {
 
     fn parse_stmt(tokens: Vec<Token>) -> Result<Stmt> {
         let mut parser = Parser::new(tokens);
-        parser.stmt()
+        parser.statement()
     }
 
     #[test]
@@ -396,34 +240,6 @@ mod tests {
                 location,
             } if location == (7,1).into()
         )
-    }
-
-    #[test]
-    fn parse_variable_declaration() {
-        let name: String = "name".into();
-        let tokens = vec![
-            token(TokenType::Var),
-            token(TokenType::Identifier(name.clone())),
-            token(TokenType::Semicolon),
-            token(TokenType::Eof),
-        ];
-        let stmts = parse_stmt(tokens).unwrap();
-        assert_eq!(stmts.to_string().trim_end(), "Var name")
-    }
-
-    #[test]
-    fn parse_variable_initialisation() {
-        let name: String = "name".into();
-        let tokens = vec![
-            token(TokenType::Var),
-            token(TokenType::Identifier(name.clone())),
-            token(TokenType::Equal),
-            token(TokenType::Nil),
-            token(TokenType::Semicolon),
-            token(TokenType::Eof),
-        ];
-        let stmts = parse_stmt(tokens).unwrap();
-        assert_eq!(stmts.to_string().trim_end(), "Var name = (nil)")
     }
 
     #[test]
@@ -551,6 +367,7 @@ mod tests {
         let stmt = parse_stmt(tokens).unwrap();
         assert_eq!(stmt.to_string().trim_end(), "while (nil) {\nExpr(nil)\n}")
     }
+
     #[test]
     fn parse_and_desugar_for() {
         let name: String = "name".into();
@@ -578,33 +395,6 @@ mod tests {
         assert_eq!(
             stmt.to_string().trim_end(),
             "{\nVar name = (nil)\nwhile (EqualEqual (variable name) (nil)) {\n{\nExpr(nil)\nExpr(name=(true))\n}\n}\n}"
-        )
-    }
-
-    #[test]
-    fn parse_function_with_2_arguments() {
-        let name: String = "name".into();
-        let arg1: String = "a".into();
-        let arg2: String = "b".into();
-
-        let tokens = vec![
-            token(TokenType::Fun),
-            token(TokenType::Identifier(name)),
-            token(TokenType::LeftParen),
-            token(TokenType::Identifier(arg1)),
-            token(TokenType::Comma),
-            token(TokenType::Identifier(arg2)),
-            token(TokenType::RightParen),
-            token(TokenType::LeftBrace),
-            token(TokenType::Nil),
-            token(TokenType::Semicolon),
-            token(TokenType::RightBrace),
-            token(TokenType::Eof),
-        ];
-        let stmt = parse_stmt(tokens).unwrap();
-        assert_eq!(
-            stmt.to_string().trim_end(),
-            "fun name(a, b, ) {\nExpr(nil)\n}"
         )
     }
 }
